+++ modules/cpdos/hbh.py
@@ -15,15 +15,16 @@ def compareRequests(url, s, headers, resp1, params2, authent):
     try:
         resp2 = s.get(url, headers=headers, params=params2, auth=authent, allow_redirects=False, verify=False, timeout=10)
     except requests.exceptions.ConnectionError as e:
-        print(f"Error : {e}")
+        #print(f"Error : {e}")
+        return None
 
-    if resp1.status_code != resp2.status_code and resp2.status_code != 429:
+    if resp1.status_code != resp2.status_code and resp2.status_code not in [429, 403] and resp1.status_code not in [301, 302]:
         behavior = "DIFFERENT STATUS-CODE"
-        print(f" \033[33m└── INTERESTING BEHAVIOR\033[0m | {VULN_NAME} | \033[34m{url}?cacheBuster={params2['cacheBuster']}\033[0m | {behavior} {resp1.status_code} > {resp2.status_code} | PAYLOAD: {headers['Connection']}")
+        print(f" \033[33m└── [INTERESTING BEHAVIOR]\033[0m | {VULN_NAME} | \033[34m{url}?cacheBuster={params2['cacheBuster']}\033[0m | {behavior} {resp1.status_code} > {resp2.status_code} | PAYLOAD: Connection: {headers['Connection']}")
         return resp2
-    if len(resp1.content) not in range(len(resp2.content) - 100, len(resp2.content) + 100) and resp2.status_code != 429:
+    if len(resp1.content) not in range(len(resp2.content) - 1000, len(resp2.content) + 1000) and resp2.status_code not in [429, 403] and resp1.status_code not in [301, 302]:
         behavior = "DIFFERENT RESPONSE LENGTH"
-        print(f" \033[33m└── INTERESTING BEHAVIOR\033[0m | {VULN_NAME} | \033[34m{url}?cacheBuster={params2['cacheBuster']}\033[0m | {behavior} {len(resp1.content)}b > {len(resp2.content)}b | PAYLOAD: {headers['Connection']}")
+        print(f" \033[33m└── [INTERESTING BEHAVIOR]\033[0m | {VULN_NAME} | \033[34m{url}?cacheBuster={params2['cacheBuster']}\033[0m | {behavior} {len(resp1.content)}b > {len(resp2.content)}b | PAYLOAD: Connection: {headers['Connection']}")
         return resp2
     return None
 
@@ -32,10 +33,11 @@ def CachePoisoning(url, s, params2, resp1, resp2, authent, headers):
     try:
         resp3 = s.get(url, params=params2, auth=authent, allow_redirects=False, verify=False, timeout=10)
     except requests.exceptions.ConnectionError as e:
-        print(f"Error : {e}")
+        pass
+        #print(f"Error : {e}")
     
     if resp3.status_code == resp2.status_code and resp3.status_code != resp1.status_code and resp2.status_code != 429:
-        print(f"  \033[31m └── VULNERABILITY CONFIRMED\033[0m | {VULN_NAME} | DIFFERENT STATUS-CODE {resp1.status_code} > {resp3.status_code} | \033[34m{url}?cacheBuster={params2['cacheBuster']}\033[0m | PAYLOAD: {headers['Connection']}")
+        print(f"  \033[31m └── [VULNERABILITY CONFIRMED]\033[0m | {VULN_NAME} | DIFFERENT STATUS-CODE {resp1.status_code} > {resp3.status_code} | \033[34m{url}?cacheBuster={params2['cacheBuster']}\033[0m | PAYLOAD: Connection {headers['Connection']}")
 
 def HBH(url, s, authent):
     uri = f"{url}?cacheBuster=coucou"
@@ -52,4 +54,5 @@ def HBH(url, s, authent):
                 if resp2:
                     CachePoisoning(url, s, params2, resp1, resp2, authent, headers)
             except Exception as e:
-                print(f"Error : {e}")
+                pass
+                #print(f"Error : {e}")